Add files via upload
Inv_Testing_v6.slx - Main program that goes between points that are listed in the program.
quick_setup_v2.mat - Sets up the variables for the main program (Inv_Testing_v6)
screen_trigger_capture_v1.py - Takes pictures using the camera based on the main program being in the hold state.
laser_centroid_calib.py - Finds the centroids of the laser pointer and calibrates the camera based on the desired coordinates. Returns a list of the laser pointer centroids and distances relative to the desired position in millimeters.